refactor(manifest): Migrate `init` method to own file
